created middleware.js file and moved middlewares to this file
