+++ middlewares.js
@@ -1,0 +1,57 @@
+const Users = require('./user-model.js');
+const bcrypt = require('bcryptjs');
+const jwt = require('jsonwebtoken');
+
+
+
+
+
+
+
+module.exports = {
+    validateUserId, authorizeUser, validateUser
+}
+
+
+async function validateUserId(req, res, next) {
+    const { userId } = req.params;
+    try {
+        const user = await Users.getUserBy({ id: userId });
+        if (user) {
+            req.user = user;
+            next();
+        } else {
+            res.status(400).json({ message: `User with Id: ${id} does not exist` })
+        }
+    }
+    catch (error) {
+        res.status(500).json({ message: "validateUserId Error", error: error })
+    }
+}
+
+const authorizeUser = (req, res, next) => {
+    const authorization = req.body.authorization;
+
+    if(authorization) {
+        jwt.verify(authorization, 'areyouforreal', (err, decodedToken) => {
+            if(err) {
+                res.status(401).json({message: 'areyouserious?'});
+            } else { 
+                req.decodedJwt = decodedToken;
+                next();
+            }
+        })
+    } else {
+        res.status(401).json({message: 'nope not gonna work'});
+    }
+};
+
+const validateUser = (req, res, next) => {
+    const { username, password } = req.body;
+
+    if (username && password ) {
+        next();
+    } else {
+        res.status(400).json({ message: 'username and password are required' });
+    }
+}
